+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Reject version bumps when there are uncommitted git changes
+
 ## [0.4.1] - 2024-06-27
 
 ### Fixed
@@ -81,7 +85,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial project setup
-
 
 [Unreleased]: https://github.com/peiman/changie/compare/0.4.1...HEAD
 [0.4.1]: https://github.com/peiman/changie/compare/0.4.0...0.4.1